python binding: initialize the xwiimote python binding
xwiimote python binding via swig supporting memory management,
exceptions, classes (monitor, event, iface). The file binding_api.txt
details the api of the binding.
The only thing which is not exactly how I wanted is the poll method
which takes an event as in-out parameter instead of just being an
output parameter, to avoid a high number of malloc.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>